backend/requirements.txt
groq
tavily-python
deepgram-sdk
python-dotenv